'database v.1.0, allowed for decimals in user_ratio table'

+++ database/migrations/2022_03_11_191657_create_user_ratio_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('user_ratio', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained('users');
-            $table->integer('ratio');
+            $table->decimal('ratio', 3, 1);
         });
     }
